Two blocks hit plane and fall asleep
I've gotten the default contact resolver and sleep episilon configured for
this scenario.  This shows how objects are intended to fall onto others
and establish resting contacts.

This is highly adapted to THIS PHYSICAL SCENARIO and it'll be part of the
development of setting up a scene to pick good iteration values.  Only
time will tell if the SleepEpsilon is gonna work for anything else.

I modified the way that the RigidBody tracks it's movement to be more
slanted towards the last value than the new one.  This means it takes more
time for the tracked motion to bleed down to the sleep threshhold

* I added a IGameEntity.Mass (getter only)
* Modified the case where two objects are colliding with inveres
  velocities, so that the resulting separating velocity is zero.
* Added a public property RigidBody.SmallPositiveMotion, which is only
  used as a debugging tool in conjunction with the entity inspector screen
* Added entity inspector attribute label to Contact.Restitution
* WhenAwakeFG now applies a mass dependent force, so it's suitable for
  gravity