Interesting case
I have changed the way that WhenAwakeFG works on boxes.  It now applies
/18th the force at each corner of the box.  I turned off the second box
and set the first box spinning on the Y axis.

When the box bounces it loses all it's Y rotation, as if the contact
resolution system thought that it was spinning on an axis that wasn't
parallel to the contact normal (the plane in this case)

TODO - Get this captured in a unit test