Guns
I added a CameraGun class and made CollideDemo have one.  The CameraGun
class renders a cross hair texture, that is red if the gun ray intersects
with a bounding sphere of an object in the ICollideables array.  It also
has a virtual Fire() method that is overriden in EntityInspectorScreenGun,
which adds an inspector screen to the game screen stack.

UserControls now has a hydrated .Firing bool, <Space> fires atm.